+++ src/weather/weatherMap.ts
@@ -2,6 +2,9 @@ import { localize } from '@/utils/game';
 import { log } from '@/utils/log';
 import { Climate, Humidity, Season } from './climateData';
 
+// drop down selections for manually setting weather
+let manualSelections: { text:string, value: string}[];
+
 // call to set everything up after the game has loaded
 function initializeLocalizedText(): void {
   log(false, 'Loading localized weather text');
@@ -37,6 +40,11 @@ function initializeLocalizedText(): void {
       }
     }
   }
+
+  // need value and text, and then a way to map the values back to the weather
+  //    for effects
+  manualSelections = manualOptions.map((val, i) =>
+      ({ value: i.toString(), text: weatherDescriptions[val?.climate as number][val?.humidity as number][val?.hexCell as number] }));
 }
 
 // rather than specifying weather by biome, we take a more flexible approach (though we also define some biomes as defaults)
@@ -392,6 +400,24 @@ const descriptionCells = {
   'winter5': 35,
   'winter9': 36,
 };
+
+const manualOptions = [  // build list of manual weather options; for simplicity, we steal them from ones that already exist
+  { climate: Climate.Cold, humidity: Humidity.Barren, hexCell: descriptionCells.springfall_cool1 },   // clear sky
+  { climate: Climate.Cold, humidity: Humidity.Modest, hexCell: descriptionCells.summer4},   // clouds
+  { climate: Climate.Cold, humidity: Humidity.Modest, hexCell: descriptionCells.winter6},   // overcast
+  { climate: Climate.Cold, humidity: Humidity.Modest, hexCell: descriptionCells.springfall_warm5},,   // dark sky
+  { climate: Climate.Cold, humidity: Humidity.Barren, hexCell: descriptionCells.summer7 },   // light rain
+  { climate: Climate.Cold, humidity: Humidity.Verdant, hexCell: descriptionCells.summer6 },   // moderate rain
+  { climate: Climate.Cold, humidity: Humidity.Modest, hexCell: descriptionCells.summer7 },   // heavy rain
+  { climate: Climate.Temperate, humidity: Humidity.Barren, hexCell: descriptionCells.springfall_cool2 },   // light fog
+  { climate: Climate.Cold, humidity: Humidity.Modest, hexCell: descriptionCells.springfall3 },   // moderate fog
+  { climate: Climate.Cold, humidity: Humidity.Modest, hexCell: descriptionCells.springfall_cool2},   // heavy fog
+  { climate: Climate.Cold, humidity: Humidity.Barren, hexCell: descriptionCells.winter5 },   // light snow
+  { climate: Climate.Temperate, humidity: Humidity.Verdant, hexCell: descriptionCells.winter9 },   // moderate snow
+  { climate: Climate.Cold , humidity: Humidity.Modest, hexCell: descriptionCells.winter3},   // heavy snow
+  { climate: Climate.Cold , humidity: Humidity.Modest, hexCell: descriptionCells.springfall_warm6},   // hail
+];
+
 
 ///////////////////////////////////////
 // define the extra metadata for each cell
@@ -871,4 +897,6 @@ export {
   weatherOptions,
   startingCells,
   nextCell,
+  manualSelections,
+  manualOptions,
 }