+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v1.5.0 - Manually select weather
+![](https://img.shields.io/badge/release%20date-October%2029%2C%202023-blue)
+![GitHub release](https://img.shields.io/github/downloads-pre/dovrosenberg/foundry-simple-weather/v1.5.0/module.zip)
+
+Enabled ability to manually select weather when needed.
+
+<hr/>
+
 ## v1.4.2 - Bug fix
 ![](https://img.shields.io/badge/release%20date-October%2028%2C%202023-blue)
 ![GitHub release](https://img.shields.io/github/downloads-pre/dovrosenberg/foundry-simple-weather/v1.4.2/module.zip)